Draft of a `GooeyUI` implementation
This is a drop-in replacement for a DataLad UI like `DialogUI` that uses
the Gooey UI for comunication with the user.

It contains a demo implementation that is able to send UI messages from
DataLad to Gooey in a thread-safe manner to the main window log viewer
widget. This is important, because
most DataLad command execution will have to take place in threads, in
order to keep the UI responsive during execution.

Closes datalad/datalad-gooey#7