Sketch for asking a question from a thread, and pass the answer back to it
This is more of a shot from the hip. The question to-be-asked is
submitted via a signal. Upon emission of the signal, the thread
waits on a get() of the message queue, until the main thread has
put in the answer given via the UI.

The whole thing is protected by a global lock (in the UI class
singleton) to make sure that only one question from a single thread can
be processed at any given time, such that the answer in the message
queue indeed belongs to that question.

Ping datalad/datalad-gooey#9 because this is the UI-half, and only the
thread-based command execution is missing now.