Rough sketch of a DataLad command configuration dialog flow
A data entry widget is generated, similar to the generation of
argparse subparsers from a command name.

Mapping of Parameter to input widgets turns out to be complicated,
because type and size information is scattered across multiple
places and often uses argparse-native implicit "side-effects".

A demo menu action is provided that populates a dialog for `wtf`.