Use QFileSystemWatcher to update visible file tree browser items
This is a key feature that makes the app react to outcomes of
external or internal processes. Two main use cases have to be
anticipated:

1. Operations being executed independently (and unknowningly)
   by some Git command/GUI. We are not planning to replace or
   wrap all Git operations, hence such independent use is to
   be expected.

2. Datalad commands being executed by the app. Being able to react
   to filessystem changes covers a large variety of command
   outcome, without the app knowing semantics of specific commands
   and there expected outcomes.

The current implementation is blunt (numerous too expensive tree(view)
updates, but outlines the concept in a semi-functional way.

Some implementations likely need to move into the DataladTreeModel
implementation.

Closes datalad/datalad-gooey#1