Much more careful tree update after `directoryChanged` signal
In order to save on resources, we do not watch every item in the
tree view with `QFileSystemWatcher`, instead we only watch expanded
directories.

This means that an incoming `directoryChanged` signal could mean a lot
of things. This change implements a more careful inspection of such
a modified directory, and performs the necessary tree item updates
in a more surgical manner.

Importantly, this should avoid the previous segfaults from forcefully
removing tree nodes indexed by tree items.

Closes datalad/datalad-gooey#27