Don't put an exception's traceback into the status bar
The former formatting of an exception from a command execution
is not well suited for a one-line status bar. It's meant for
printing a human-readable multi-line traceback.
Change to a short formatting that gives just the exception class
and its message.

Note, that a traceback woud show up at loglevel 8 anyway via
CapturedException's constructor.