Insert command group submenues in right order
The dataset method menu shows submenu items corresponding to dataset
method groups in an unintuitive order - not whats most useful is at the
top, but the group who the first command in an alphabetically sorted
list belongs to.  From a UX standpoint, the first group should be the
most essential commands.
A sensible order of command groups already exists in the menu_lookup,
but this order is lost during menu assembly, as the assembly proceeds by
alphabetical order of command names (see #34). This change preserves the
original order (by adding the submenues in the order they are supposed
to be) after each submenu was populated with commands.