Add common opts to command tab
This is limited (for now) to `on_failure` and `result_renderer` as the
other ones don't seem to make sense in the GUI at this point. Eventually
`return_type` may become relevant as well.

This also comes with a slight change to the parameter widget helpers:
Passing down `Parameter` instances instead of the command class, since
this is the only thing the command class was used for and this enables
reuse of the helpers outside the context of a specific command.
Also: `allargs` was declared to possibly be None by an intermediate
helper, who would then pass it down to another one where `allargs` was
mandatory. Allow for `None` all the way now.