Speed-up status annotation (massively) by limiting to subds commits
This prevents subdatasets from being inspected for untracked content
(even when the respective directories are not yet expanded).

Effectively, a subdataset annotation (attached to the root directory
of a subdataset within a superdataset) now reflects whether the
state of the subdataset recorded _within_ the superdataset matches
the subdatasets checked-out state, or not.