Speed up item lookup in tree widget
Before we searched linearily through the list of items.
On my machine, this took between 1ms and 100ms per search for
a list of 36k children -- prohibitively slow.

The change introduces a lookup dictionary in each item. This brings
down lookup time to 1ms per item, even in lists of >30k items.

This is a reasonable upper limit for a "typical" directory. A complete
lookup (e.g. for re-annotation) would take ~30s. Not snappy, but still
tolerable IMHO.

This now comes with the need to maintain that lookup, but the costs
are outweighed by the benefits.