Unify command execution Signals
There are two Signals triggering a datalad command execution:
`configure_dataladcmd` and `execute_dataladcmd`. Definition of the
former did not match the latter resulting in `GooeyDataladCmdUI`
emmitting a signal that the receiver did not properly handle, b/c it got
a `None` where it expects a `dict` leading to an "invisible"
`AttributeError`.

(Closes #84)