Implement fast version checking of current DataLad against latest on Pypi.org
This fixes #81.
Internally, it relies on a small Python library (outdated), which performs
plain HTTP requests against Pypi.org for the latest version of a specific
software package. This makes the version check independent of the presence
of pip locally, and also much faster.
The outcome of the check is written into the console log.