Make user pick a directory when no path is given
This change adds a FileDialog that prompts users to pick a directory
on their system before showing the App. When the user picks a directory,
it is given as a path to the App. When the user presses 'Cancel', an empty
path is passed to the App, which defaults to the current behavior of opening
the App in PWD. When the gooey is invoked with the -p path flag, the
FileDialog doesn't show up.