API override specification
This is a system that can be used to declare a specific API subset

One can set

- which commands are included
- which labels command are presented with in the UI
- exclusion of parameters by name
  - across all commands
  - for individual commands
- display names for parameters
  - generic across all commands
  - specific for individual commands
- a specific order of parameters in the configuration form
  (deviating from a default order by parameter name)

Multiple API specifications of this kind can be maintained.
A central "active_api" switch is provided to not translate this
complexity into all code.

A demo is provided for the `create` command in the `simplified` API.