Simplistic approach to stop a command execution thread
Whenever a command execution starts, a tool button is shown in the
status bar. When clicked that button causes a stop flag to be set in
the `GooeyDataladCmdExec`, which causes the execution thread to
raise an `InterruptedError` exception when it receives the next command
result.

All gather results (up to this point) are captured and emitted.
The resulting code path is the same as if the command had crashed on its
own.

On exit (forced or natural), the kill button is hidding again.

Right now the implementation can only deal with a single running thread,
so it needs to be adjusted, whenever we start supporting more than one.

We likely need to support requesting a manual re-exploration of
directories in the file browser now, because user can (and will)
interrupt the initial traversal and annotation runs too. And the need to
be able to re-request them, after a moment of regret.

Closes datalad/datalad-gooey#31