Add a .desktop file
File contents proposed by @mih and included without changing. Reasons
for choosing the location below.

Since we are already grafting `datalad_gooey` in `MANIFEST.in`, this
file will be automatically included with package files. From there, it
can be moved to a proper place after installation. Placing it on the
project's top-level and explicitly including in `MANIFEST.in` would
include the file in the source package, but not make it part of the
package directory (site-packages) after installation.

Moving this file is left for after installation because, apparently,
doing so is out of the intended scope of `pip install` -- writing to
platform-specific locations external to the python package is better
left to OS-level installers (or done explicitly by users). The
`data_files` argument to `setup` (in `setup.py`), which allowed to
move data files to external locations, has been deprecated.