Use Python package name of dark theme dependency
It tried out the dark theme, and wanted to follow the warnings advice of
installing the 'missing qdarktheme installation'. However, pip install
qdarktheme does not find a matching package, and I didn't find a different
package manager that contained a package of this name. I suppose that the
relevant Python package is called pyqtdarktheme, and I think its easier
to find this out if the warning uses its package name