Make GooeyApp set PWD to match its root path
This makes any command run in the root that is visible in the app,
rather than some inaccessible (likely not useful) working directory.

Now one could run a plain `create()` and not get an wholly unexpected
result.

Closes datalad/datalad-gooey#168

Moreover, any relative path is not relative to the root item in the
treeview. Which is better than relative-to-who-knows.

Ping datalad/datalad-gooey#106

This change paves the way to an implementation of fix for
datalad/datalad-gooey#130