Sibling selection for `push` and `update`
This demos how we can perform more detailed type-checking, and
select more targeted input widgets based on that.

The API specifications now support `parameter_constraints` dicts
that take  `Constraint` instances that replace those in the
original parameter specification.

A demo `EnsureDatasetSiblingName` is provided, and applied to
`update(sibling=)` and `push(to=)`.

The constraint type is used to select the new `SiblingChoiceParamWidget`
which pulls the list of siblings from a dataset, once it is known.

Closes datalad/datalad-gooey#172