Fix defaults for parameter widgets
This adds a test for parameter widgets to deliver a representation of
the default value (currently `_NoValue`) after initialization with the
parameter specification and fixes several widget classes that failed
this test, delivering the native default of the respective Qt widget
instead.
It also fixes the Mixin implementation of `get_gooey_param_spec` to
deliver the same representation of default under all conditions.

Alternatively, `ChoiceParamWidget` and `PosIntParamWidget` could
implement an equivalent of the `StrParamWidget`'s `setModified` approach
by wiring up the respective signals, but this simpler approach of just
setting the default as the current value also serves the purpose of a
visual representation of the default for those widget types.