Minimal progress reporting in the status bar
This is not the epic implementation of datalad core, no units, no 12
progress bars.

There is one progress bar whenever there is any progress reporting
ongoing.

If there is more than one process reporting progress at a time, the
"average progress" is reported. This makes the progress a bit jumpy.

If a particular progress tracker does not have a known total value
(aspirational goal), it is presently ignored. However, the Qt
widget allows for "busy indicator" mode. So this could be supported top.

Closes datalad/datalad-gooey#156