New MultiValueInputWidget implementation
It now longer does the previous in-item editing approach. Instead,
it features a persistent editor. It should no longer crash.

Closes datalad/datalad-gooey#114

This approach could be extended to support multiple input widgets for
different types that joint feed into a list of items. This would be a
possible solution to datalad/datalad-gooey#90