Multiline text edit widget for message param
Provide a TextEdit widget and associate it with the `message` parameter.
This widget allows for drag&drop. However, currently it's getting the
file URL of a local text file rather than its content. This should
change.
Furthermore, disabled RichText/HTML for now to keep it simple until this
is better tested.

Addresses #93