Give user explicit means to refresh the tree browser
This should not be necessary, but maybe there are corner cases left,
where it is needed, e.g. https://github.com/datalad/datalad-gooey/issues/239,
or a user managed to KABOOM the tree population itself.

For now this is the simplest possible implementation (reload the whole
thing). Later this could become more clever and more targeted.

Closes #239