Extend mixin API with helepr for drag&drop event handling
This generalizes the previous implementation, and extends it to
`MultiValueInputWidget`. This meta widget will not pass drop events
onto an underlying editor widget, if that supports an event.

For unsupported events that comprise of URLs (e.g. `file://`
URLs like one would get from a file manager), it passes them
on, one-by-one, to the editor, and then builds new items
from the editor representation of that URL (e.g. a path).

Ping #190