Store/restore "main" geometry/state - window and the file tree header
It felt inconvenient to need to rescale header (columns) in the
tree view each time and/or make window bigger.  I guess there could be
more geometries and states to store/restore but this is the main
ones I felt need for.

I guess such functionality could become configurable, but as IMHO should
be enabled by default for better UX decided to suggest hardcoded first.