Rewrite Parameter handling
This change replaces the previously used mixin pattern for parameter
input widget. Now the Gooey parameter representation and the input
widget are two connected but separate entities.

`GooeyCommandParameter` is the new base class that encapsulates all
parameter value handling that is not internal to a single input widget.

- Parameter name and default handling
- Reporting of values, only when they are different from the default
- Only accepting values that pass validation with an assigned constraint
- Tailoring constraints to datasets-contexts (and doing that, if needed)
- Qt input widget generation
- Qt label generation for displaying parameter names
- Presenting parameter documentation as tooltips
- Drag&drop support helpers

The validator/constraint integration is improved. Any parameter value
change will now trigger a validation run of the entire command
configuration form. Invalid parameter values are flagged with a visual
indicator next to the parameter name. As long as invalid parameter
values exist, command submission is prevented.  Because this is done
immediately, and continuously, users also get an immediate indication
of require command parameters -- something that required trail-and-error
before.

- Closes #290
- Closes #287