Rethink parameter widget selection
Before it was a mess, now it is less of a mess.

Importantly, the parameter widget selection is no longer based on a
mixture of parameter names, constraints, and argparse internal
settings. Now the selection is made solely based on an assigned
`Constraint`.

To make this possible the whole slew of new Constraints that came to
life lately is put to action. The constraint provided by -core's
`Parameter` class is now treated as an "item constraint". This is
turned into a "comprehensive constraint" by wrapping it, as necessary,
into additional constraints. E.g. `nargs='*'` causes
`EnsureIterableOf(list, constraint)`.

A number of other rules are implemented to always yield a single
constraint as a fully qualified ruleset to translate a parameters
requirements into a widget(set).

Importantly, it is not necessary to indicate an explicit `|
EnsureNone()` for a parameter that takes totally different values, but
defaults to `None`, merely to indicate that it is optional.
Now, an `EnsureNone()` is added automatically.

The approach should be extended to an `EnsureValue(default)`, for any
default value that is not already covered by the assigned constraint.

Closes datalad/datalad-gooey#302