Basic setup for a "Metadata" context tab
A handler widget (`MetadataWidget`) is permanently hosted in that tab.
It provides path and metadata type selection services. Metadata-type
specific editor (all derived from a common `MetadataEditor` base class
hook into it, by being dynamically loaded, based on the specified type.

A private `GooeyApp` slot `_edit_metadata()` is provided that pulls
up the editor, and can be used to connect menu actions to.
A demo menu action is added for all tree items of type `annexed-file`
that triggers an annex metadata editor (not yet functional) to be opened.