Stop unwatching directories when tree item is collapsed
This behavior has a number of issues. It stops monitoring that
particular directory, but not any of the subdirectories that might
also be watched (although equally invisible).

Moreover, re-expanding such an item will put it back on the watch list,
but it will not repopulate the item with any changes that occurred in
in the meantime.

Overall, this leads to inconsistencies between FS and tree view.

The original intention was to save on resources for watching. However,
even on the most limited systems a few dozens are easily possible.
I'd say we tune this whenever we actually hit a use case where this
is a problem.

Until then, we never stop watch a directory that we started watching
once. `QFileSystemWatcher` is already taking care of automatically
unwatching directories that get deleted.

- Closes datalad/datalad-gooey#50