Prevent context menu actions to run on absent datasets
This is the core issue of datalad/datalad-gooey#336

Here we change the dataset context to the superdataset, and declare the
clicked dataset as the `path`. This should be universally safe.
The only dataset that could be "absent" is a SUBdataset.

This could theroretically cause issues with command that take no
`path` arguments. However, it is likely also not sensible to expose
such commands for absent subdataset. If this turns our to be a real
issue the fix should be along these lines.

This change, however, is not sufficient to address datalad/datalad-gooey#336
completely, because actually getting/cloning a subdataset also requires
setting `--recursive`, but we likely do not want to set it without a
limit -- which is not exposed in the simplified API.

For that we could address datalad/datalad-gooey#219, and set
`recursion_limit=1` for `get`.