Annotate symlinks with their targets
Until now, symlinks were just labeled as such with no indication of
their nature.

This change annotates symlinks, both in terms of `lsdir` results and
their representation in the FS tree, with their targets.

While a user still cannot "follow" them within the Gooey, they can at
least see what is going on.

Before `status` annotation, also annex'ed files are marked up in this
fashion, because I see no need to make an expensive special case for
that.

Closes datalad/datalad-gooey#23

For the future, one could think about supporting a double-click event
on a symlink pointing to a directory that is also in the tree view,
and switching to the respective item as current.