Double-click on a symlink jumps to (closest) target
It will expand and scroll the tree to the target item, if it
exists. If not, it will go to the closest existing location.
It remains to be seen whether that is a universally appreciated
behavior. But a force exploration of the tree could take a substantial
amount of time.