use datalad-gooey-askpass in Windows
This commit enables the windows version of datalad-gooey
to interact with users graphically, using `datalad-gooey-askpass`.

It sets the environment variable `$env:DISPLAY`, which seems to
be necessary in Windows to trigger the use of the asker that is
stored in `$env:SSH_ASKPASS`, even if `$env:SSH_ASKPASS_REQUIRE`
is to `force`.

This change is done in the starter script of datalad-gooey in order
to not mess with $DISPLAY on X11-like systems.

This fixes issue #371