Stop caching `FSBrowserItem._get_item_from_path()`
The cache was not invalidated properly, leading to deleted items beeing
handed out. This is behind a whole range of ill behavior (likely).

Re-assessing the performance impact of caching, I found it to be
neglible. Hence I disabled it with a comment, instead of implementing
cache invalidation (which is fairly doable, but seems not useful).

Closes datalad/datalad-gooey#400
Closes datalad/datalad-gooey#399
Closes datalad/datalad-gooey#398
Closes datalad/datalad-gooey#397
Closes datalad/datalad-gooey#396
Closes datalad/datalad-gooey#395
Closes datalad/datalad-gooey#387
Closes datalad/datalad-gooey#379