DOC: Update launch instructions for Mac
Currently we don't have a method for putting gooey into the
Applications folder, and recommend `pip install` on MacOS. This means
that the way to launch gooey on MacOS is through the
terminal. Suggesting using Applications to launch could cause user
confusion.

This will need to be reverted / updated if we add a way to include
gooey in Applications.